+++ packages/building-rollup/modern-config.js
@@ -1,5 +1,6 @@
 // @ts-nocheck
 
+const { findSupportedBrowsers } = require('@open-wc/building-utils');
 const resolve = require('rollup-plugin-node-resolve');
 const { terser } = require('rollup-plugin-terser');
 const babel = require('rollup-plugin-babel');
@@ -30,7 +31,7 @@ module.exports = function createBasicConfig(_options) {
         }),
 
       // parse input index.html as input and feed any modules found to rollup
-      modernWeb(),
+      options.input.endsWith('.html') && modernWeb(),
 
       // resolve bare import specifiers
       resolve(),
@@ -48,14 +49,7 @@ module.exports = function createBasicConfig(_options) {
           [
             '@babel/env',
             {
-              targets: [
-                'last 2 Chrome major versions',
-                'last 2 ChromeAndroid major versions',
-                'last 2 Edge major versions',
-                'last 2 Firefox major versions',
-                'last 2 Safari major versions',
-                'last 2 iOS major versions',
-              ],
+              targets: findSupportedBrowsers(),
               useBuiltIns: false,
             },
           ],