+++ packages/building-webpack/modern-config.js
@@ -1,3 +1,4 @@
+const { findSupportedBrowsers } = require('@open-wc/building-utils');
 const path = require('path');
 const HtmlWebpackPlugin = require('html-webpack-plugin');
 const CleanWebpackPlugin = require('clean-webpack-plugin');
@@ -94,7 +95,7 @@ module.exports = userOptions => {
                 ['bundled-import-meta', { importStyle: 'baseURI' }],
               ].filter(_ => !!_),
 
-              presets: ['@babel/preset-env'],
+              presets: [['@babel/preset-env', { targets: findSupportedBrowsers() }]],
             },
           },
         },