+++ packages/create/CHANGELOG.md
@@ -3,6 +3,47 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.11.0](https://github.com/open-wc/open-wc/compare/@open-wc/create@0.10.1...@open-wc/create@0.11.0) (2019-05-07)
+
+
+### Features
+
+* **create:** remove browerslistrc ([1930660](https://github.com/open-wc/open-wc/commit/1930660))
+
+
+
+
+
+## [0.10.1](https://github.com/open-wc/open-wc/compare/@open-wc/create@0.10.0...@open-wc/create@0.10.1) (2019-05-06)
+
+**Note:** Version bump only for package @open-wc/create
+
+
+
+
+
+# [0.10.0](https://github.com/open-wc/open-wc/compare/@open-wc/create@0.9.5...@open-wc/create@0.10.0) (2019-05-06)
+
+
+### Features
+
+* update to latest testing-karma config syntax ([465bfe0](https://github.com/open-wc/open-wc/commit/465bfe0))
+
+
+
+
+
+## [0.9.5](https://github.com/open-wc/open-wc/compare/@open-wc/create@0.9.4...@open-wc/create@0.9.5) (2019-05-03)
+
+
+### Bug Fixes
+
+* **create:** fix tests in barebone-app ([b6a88cc](https://github.com/open-wc/open-wc/commit/b6a88cc))
+
+
+
+
+
 ## [0.9.4](https://github.com/open-wc/open-wc/compare/@open-wc/create@0.9.3...@open-wc/create@0.9.4) (2019-04-28)